Refactor to improve remote logic and performance

+++ lib/services/shell_service.dart
@@ -1,14 +1,28 @@
+import 'dart:async';
 import 'dart:io';
 
 import 'package:path/path.dart' as path;
 import 'package:process_run/process_run.dart';
 
 import '../exceptions/file_system_exceptions.dart';
+import '../models/io_models.dart';
 
 class ShellService {
   Shell _shell;
+  int? _pid;
+  final bool _isWindows;
+  final ProcessRunSync _runProcessSync;
+  final ProcessSignal _processSignal;
 
-  ShellService({Shell? shell}) : _shell = shell ?? Shell(verbose: false);
+  ShellService({
+    Shell? shell,
+    bool? isWindows,
+    ProcessRunSync? runProcessSync,
+    ProcessSignal? processSignal,
+  })  : _shell = shell ?? Shell(verbose: false),
+        _isWindows = isWindows ?? Platform.isWindows,
+        _runProcessSync = runProcessSync ?? Process.runSync,
+        _processSignal = processSignal ?? ProcessSignal.sigint;
 
   void checkExecutable(String executable) {
     final execLocation = path.basename(executable) == executable
@@ -21,7 +35,16 @@ class ShellService {
   }
 
   Future<List<ProcessResult>> runScript(String script) {
-    return _shell.run(script);
+    return _shell.run(script, onProcess: (process) {
+      if (_pid == null) {
+        _pid = process.pid;
+        late final StreamSubscription<ProcessSignal> sub;
+        sub = _processSignal.watch().listen((event) {
+          sub.cancel();
+          dispose();
+        });
+      }
+    });
   }
 
   void moveShellTo(String path) {
@@ -33,6 +56,10 @@ class ShellService {
   }
 
   void dispose() {
+    if (_pid != null && _isWindows) {
+      _runProcessSync('taskkill', ['/pid', '$_pid', '/f', '/t']);
+      return;
+    }
     _shell.kill();
   }
